Add actions module to formalise message passing
Specifically, there is an ActionsMixin now which you can invoke
with a list of action names to listen for. When these actions
occur, the mixin will dump the payload.data to a state variable
of the same name of the action. This drastically cuts down on
boilerplate code for registering dispatchers, listening for actions
and dumping some state into this.state

Since it's now all centralised we also do basic schema type checking
on emitted actions, making sure that 1) actions are registered in
actions.js and 2) that the emitted action has the keys of the right
types in the .data property.