Add Controller and GithubApi classes
The Controller class is the main entity which orchestrates the UI
and the other logic classes. The GithubApi is a wrapper around
Github HTTP requests. All of this is tied together in app.js which
forms the dependency graph for logic classes (NB: the dependency
graph for UI components is more flexible as they are dumped via
props, which makes sense when you're dealing with UIs which change
frequently).