Assume the last line in the patch is the comment position
Github's API exposes an 'original_position' which lies, resulting
in missing line comments since it doesn't know where to put them.
The diff hunk supplied in the same API is better as it gets the
right position (when you make the assumption that the last line
on the hunk is the comment position).