Refactor comment box logic
Make CommentBox dumb and just call onSubmit on its' prop, which
bubbles up from CommentListView to PullRequestOverview which gives
it context (overview comment; the pr) and bubbles it up to MainPage
which will now be the gatekeeper to invoking the Controller *directly*
rather than via actions indirection. This is desirable for things
which really make sense to have Promise-style req/responses (posting
comments) but the actions stuff is still relevant for the reverse
(e.g. refreshing the page after the comment is posted).

THIS BREAKS LINE COMMENTS.