Bug fix for line comments
Sometimes the head sha in the PR response is different from the
sha of the last commit in the /commits API. I don't know why this
may be, but it can lead to failures when leaving line comments
e.g. because the file didn't exist at this earlier commit so you
get an invalid path error. Given I thought they were equivalent,
this should hopefully not introduce regressions :D