Compensate for possibly-empty commit users
If a commit was made by an author whose email address isn't linked to a
github account, there is no author object in the commit.

There is, however, a commit object in the commit, which contains a
(different) author object with non-account-related metadata. Grab some
data from there instead.

This was found when NegativeMjark tried to use review.rocks, and it
turns out one of his commits from one of his PRs was not associated with
an account. He conveniently provided me with a har file which made
tracking this down pretty easy. I've added that example PR (its first
commit is lacking an author) to examples, as well as a script for
extracting github responses from har files, because I'm sure it will be
useful again.