Fix line diff anchors. Allow personal access tokens on login
Line diff anchors: They assumed that the from SHA was always BASE
which isn't true. They now have the form
"diff-$from-$to-$filehash$linenum".
Tweak comment CSS to set a fixed-height header. Previously the height
was determined by the size of the image buttons but this meant that
anchors would be a little out if you jumped to the anchor and THEN
the image loaded.

Personal access tokens: Add a box to specify a personal access token
on the login page so you can use the app without having a hard-dep
on review.rocks (especially since it doesn't work on localhost urls!).